Add a script to compile Apache Math
GenProg uses javac by default, but it doesn't work for Math because of
external dependencies. Instead, we use a custom Python script to move
mutated files in and out of the src folder, compile the mutated code
by calling ant, and revert the src folder back to the original state.

When setting up new Math projects, create a symlink to this script
should be sufficient.